requirements.txt
boto3==1.9.253
botocore==1.16.19
Brlapi==0.7.0
cached-property==1.5.1
cachetools==4.1.0
certifi==2019.11.28
chardet==3.0.4
Click==7.0
colorama==0.4.3
command-not-found==0.3
cryptography==2.8
cupshelpers==1.0
cycler==0.10.0
Cython==0.29.14
dbus-python==1.2.16
decorator==4.4.2
defer==1.0.6
distro==1.4.0
distro-info===0.23ubuntu1
dnspython==1.16.0
docker==4.1.0
docker-compose==1.25.0
dockerpty==0.4.1
docopt==0.6.2
docutils==0.16
duplicity==0.8.12.0
entrypoints==0.3
fasteners==0.14.1
flake8==3.7.9
Flask==1.1.1
future==0.18.2
gast==0.3.3
google-api-core==1.17.0
google-api-python-client==1.8.3
googleapis-common-protos==1.51.0
google-auth==1.15.0
google-auth-httplib2==0.0.3
google-auth-oauthlib==0.4.1
google-pasta==0.2.0
grpcio==1.29.0
h5py==2.10.0
httplib2==0.18.1
idna==2.8
importlib-metadata==1.5.0
imutils==0.5.3
ipython==7.13.0
ipython-genutils==0.2.0
isort==4.3.4
itsdangerous==1.1.0
jedi==0.15.2
Jinja2==2.10.1
jmespath==0.9.4
joblib==0.14.0
jsonschema==3.2.0
Keras==2.2.4
Keras-Applications==1.0.6
Keras-Preprocessing==1.1.2
keyring==18.0.1
kiwisolver==1.0.1
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
lazy-object-proxy==1.4.3
ldap3==2.4.1
lockfile==0.12.2
logilab-common==1.4.3
louis==3.12.0
macaroonbakery==1.3.1
Mako==1.1.0
Markdown==3.2.2
MarkupSafe==1.1.0
matplotlib==3.1.2
mccabe==0.6.1
monotonic==1.5
more-itertools==4.2.0
mpi4py==3.0.3
netaddr==0.7.19
netifaces==0.10.4
nose==1.3.7
ntlm-auth==1.1.0
numpy==1.17.4
oauth2client==4.1.2
oauthlib==3.1.0
olefile==0.46
onboard==1.4.1
openshot-qt==2.4.3
opt-einsum==3.2.1
packaging==20.3
parameterized==0.7.0
paramiko==2.6.0
parso==0.5.2
pbr==5.4.5
pexpect==4.6.0
pickleshare==0.7.5
piecash==1.0.0
Pillow==7.0.0
pluggy==0.13.0
prompt-toolkit==2.0.10
protobuf==3.12.2
psutil==5.5.1
py==1.8.1
pyasn1==0.4.2
pyasn1-modules==0.2.1
PyAudio==0.2.11
pycairo==1.16.2
pycodestyle==2.5.0
pycrypto==2.6.1
pycups==1.9.73
pydot==1.4.1
pyeclib==1.6.0
pyflakes==2.1.1
pygame==1.9.6
Pygments==2.3.1
PyGObject==3.36.0
pyinotify==0.9.6
PyJWT==1.7.1
pykerberos==1.1.14
pylint==2.4.4
pymacaroons==0.13.0
PyNaCl==1.3.0
pyOpenSSL==19.0.0
pyparsing==2.4.6
PyQt5==5.14.1
pyRFC3339==1.1
pyrsistent==0.15.5
pyserial==3.4
pytest==4.6.9
python-apt==2.0.0+ubuntu0.20.4.1
python-dateutil==2.7.3
python-debian===0.1.36ubuntu1
python-dotenv==0.13.0
python-k8sclient==0.4.0
pytz==2019.3
pywinrm==0.3.0
PyX==0.15
pyxdg==0.26
PyYAML==5.3.1
pyzmq==18.1.1
redis==3.3.11
reportlab==3.5.34
requests==2.22.0
requests-kerberos==0.12.0
requests-ntlm==1.1.0
requests-oauthlib==1.3.0
requests-unixsocket==0.2.0
roman==2.0.0
rsa==4.0
s3transfer==0.3.3
scapy==2.4.3
scikit-learn==0.22.2.post1
scipy==1.4.1
SecretStorage==2.3.1
selinux==3.0
simplejson==3.16.0
sip==4.19.21
six==1.14.0
specan==0.0.0
SQLAlchemy==1.3.12
SQLAlchemy-Utils==0.36.6
stravalib==0.10.2
systemd-python==234
system-service==0.3
tensorboard==2.2.2
tensorboard-plugin-wit==1.6.0.post3
tensorflow==2.2.0
tensorflow-estimator==2.2.0
termcolor==1.1.0
texttable==1.6.2
Theano==1.0.4
traitlets==4.3.3
tzlocal==2.1
ubuntu-advantage-tools==20.3
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
units==0.7
uritemplate==3.0.1
urllib3==1.25.8
usb-creator==0.3.7
vboxapi==1.0
wadllib==1.3.3
wcwidth==0.1.8
websocket-client==0.53.0
webstorageS3==1.0.0
Werkzeug==0.16.1
wrapt==1.11.2
xdiagnose==3.8.9
xkit==0.0.0
xmltodict==0.12.0
zipp==1.0.0
